add bff template risk analysis route
