+++ packages/eservice-template-process/src/services/eserviceTemplateService.ts
@@ -5,10 +5,13 @@ import {
   AuthData,
   DB,
   FileManager,
+  RiskAnalysisValidatedForm,
   WithLogger,
   eventRepository,
   hasPermission,
   userRoles,
+  riskAnalysisValidatedFormToNewRiskAnalysis,
+  validateRiskAnalysis,
 } from "pagopa-interop-commons";
 import {
   EServiceTemplate,
@@ -18,15 +21,26 @@ import {
   EServiceTemplateVersionId,
   EServiceTemplateVersionState,
   eserviceTemplateVersionState,
+  generateId,
+  RiskAnalysis,
+  RiskAnalysisId,
+  Tenant,
+  TenantId,
+  TenantKind,
   WithMetadata,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
+import { eserviceTemplateApi } from "pagopa-interop-api-clients";
 import {
+  eserviceTemaplateRiskAnalysisNameDuplicate,
   eServiceTemplateDuplicate,
   eServiceTemplateNotFound,
+  eserviceTemplateRequesterIsNotCreator,
   eServiceTemplateVersionNotFound,
   eserviceTemplateWithoutPublishedVersion,
   notValidEServiceTemplateVersionState,
+  riskAnalysisValidationFailed,
+  tenantNotFound,
 } from "../model/domain/errors.js";
 import {
   toCreateEventEServiceTemplateAudienceDescriptionUpdated,
@@ -34,9 +48,17 @@ import {
   toCreateEventEServiceTemplateVersionActivated,
   toCreateEventEServiceTemplateVersionSuspended,
   toCreateEventEServiceTemplateNameUpdated,
+  toCreateEventEServiceTemplateRiskAnalysisAdded,
+  toCreateEventEServiceTemplateRiskAnalysisDeleted,
+  toCreateEventEServiceTemplateRiskAnalysisUpdated,
 } from "../model/domain/toEvent.js";
 import { ReadModelService } from "./readModelService.js";
-import { assertRequesterEServiceTemplateCreator } from "./validators.js";
+import {
+  assertIsDraftTemplate,
+  assertIsReceiveTemplate,
+  assertRequesterEServiceTemplateCreator,
+  assertTenantKindExists,
+} from "./validators.js";
 
 export const retrieveEServiceTemplate = async (
   eserviceTemplateId: EServiceTemplateId,
@@ -154,6 +176,29 @@ const replaceEServiceTemplateVersion = (
     versions: updatedEServiceTemplateVersions,
   };
 };
+
+const retrieveTenant = async (
+  tenantId: TenantId,
+  readModelService: ReadModelService
+): Promise<Tenant> => {
+  const tenant = await readModelService.getTenantById(tenantId);
+  if (tenant === undefined) {
+    throw tenantNotFound(tenantId);
+  }
+  return tenant;
+};
+
+export function validateRiskAnalysisSchemaOrThrow(
+  riskAnalysisForm: eserviceTemplateApi.EServiceRiskAnalysisSeed["riskAnalysisForm"],
+  tenantKind: TenantKind
+): RiskAnalysisValidatedForm {
+  const result = validateRiskAnalysis(riskAnalysisForm, true, tenantKind);
+  if (result.type === "invalid") {
+    throw riskAnalysisValidationFailed(result.issues);
+  } else {
+    return result.value;
+  }
+}
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function eserviceTemplateServiceBuilder(
@@ -416,6 +461,155 @@ export function eserviceTemplateServiceBuilder(
       );
 
       return applyVisibilityToEServiceTemplate(eserviceTemplate.data, authData);
+    },
+
+    async createRiskAnalysis(
+      id: EServiceTemplateId,
+      createRiskAnalysis: eserviceTemplateApi.EServiceRiskAnalysisSeed,
+      { authData, correlationId, logger }: WithLogger<AppContext>
+    ): Promise<void> {
+      logger.info(`Creating risk analysis for eServiceTemplateId: ${id}`);
+
+      const template = await retrieveEServiceTemplate(id, readModelService);
+      if (template.data.creatorId !== authData.organizationId) {
+        throw eserviceTemplateRequesterIsNotCreator(id);
+      }
+      assertIsDraftTemplate(template.data);
+      assertIsReceiveTemplate(template.data);
+
+      const tenant = await retrieveTenant(
+        template.data.creatorId,
+        readModelService
+      );
+      assertTenantKindExists(tenant);
+
+      const raSameName = template.data.riskAnalysis.find(
+        (ra) => ra.name === createRiskAnalysis.name
+      );
+      if (raSameName) {
+        throw eserviceTemaplateRiskAnalysisNameDuplicate(
+          createRiskAnalysis.name
+        );
+      }
+
+      const validatedRiskAnalysisForm = validateRiskAnalysisSchemaOrThrow(
+        createRiskAnalysis.riskAnalysisForm,
+        tenant.kind
+      );
+
+      const newRiskAnalysis: RiskAnalysis =
+        riskAnalysisValidatedFormToNewRiskAnalysis(
+          validatedRiskAnalysisForm,
+          createRiskAnalysis.name
+        );
+
+      const newTemplate: EServiceTemplate = {
+        ...template.data,
+        riskAnalysis: [...template.data.riskAnalysis, newRiskAnalysis],
+      };
+
+      const event = toCreateEventEServiceTemplateRiskAnalysisAdded(
+        template.data.id,
+        template.metadata.version,
+        generateId<RiskAnalysisId>(),
+        newTemplate,
+        correlationId
+      );
+
+      await repository.createEvent(event);
+    },
+    async deleteRiskAnalysis(
+      templateId: EServiceTemplateId,
+      riskAnalysisId: RiskAnalysisId,
+      { authData, correlationId, logger }: WithLogger<AppContext>
+    ): Promise<void> {
+      logger.info(
+        `Deleting risk analysis with id: ${riskAnalysisId} from eServiceTemplate with id: ${templateId}`
+      );
+
+      const template = await retrieveEServiceTemplate(
+        templateId,
+        readModelService
+      );
+      if (template.data.creatorId !== authData.organizationId) {
+        throw eserviceTemplateRequesterIsNotCreator(templateId);
+      }
+      assertIsDraftTemplate(template.data);
+      assertIsReceiveTemplate(template.data);
+
+      const newTemplate: EServiceTemplate = {
+        ...template.data,
+        riskAnalysis: template.data.riskAnalysis.filter(
+          (ra) => ra.id !== riskAnalysisId
+        ),
+      };
+
+      const event = toCreateEventEServiceTemplateRiskAnalysisDeleted(
+        template.data.id,
+        template.metadata.version,
+        riskAnalysisId,
+        newTemplate,
+        correlationId
+      );
+
+      await repository.createEvent(event);
+    },
+    async updateRiskAnalysis(
+      templateId: EServiceTemplateId,
+      riskAnalysisId: RiskAnalysisId,
+      updateRiskAnalysis: eserviceTemplateApi.EServiceRiskAnalysisSeed,
+      { authData, correlationId, logger }: WithLogger<AppContext>
+    ): Promise<void> {
+      logger.info(
+        `Updating risk analysis with id: ${riskAnalysisId} from eServiceTemplate with id: ${templateId}`
+      );
+
+      const template = await retrieveEServiceTemplate(
+        templateId,
+        readModelService
+      );
+      if (template.data.creatorId !== authData.organizationId) {
+        throw eserviceTemplateRequesterIsNotCreator(templateId);
+      }
+      assertIsDraftTemplate(template.data);
+      assertIsReceiveTemplate(template.data);
+
+      const tenant = await retrieveTenant(
+        template.data.creatorId,
+        readModelService
+      );
+      assertTenantKindExists(tenant);
+
+      const validatedRiskAnalysisForm = validateRiskAnalysisSchemaOrThrow(
+        updateRiskAnalysis.riskAnalysisForm,
+        tenant.kind
+      );
+
+      const updatedRiskAnalysis: RiskAnalysis =
+        riskAnalysisValidatedFormToNewRiskAnalysis(
+          validatedRiskAnalysisForm,
+          updateRiskAnalysis.name
+        );
+
+      const newTemplate: EServiceTemplate = {
+        ...template.data,
+        riskAnalysis: [
+          ...template.data.riskAnalysis.filter(
+            (ra) => ra.id !== riskAnalysisId
+          ),
+          updatedRiskAnalysis,
+        ],
+      };
+
+      const event = toCreateEventEServiceTemplateRiskAnalysisUpdated(
+        template.data.id,
+        template.metadata.version,
+        riskAnalysisId,
+        newTemplate,
+        correlationId
+      );
+
+      await repository.createEvent(event);
     },
   };
 }
