+++ packages/eservice-template-process/test/utils.ts
@@ -20,7 +20,6 @@ import { eserviceTemplateApi } from "pagopa-interop-api-clients";
 import { riskAnalysisFormToRiskAnalysisFormToValidate } from "pagopa-interop-commons";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { eserviceTemplateServiceBuilder } from "../src/services/eserviceTemplateService.js";
-import { riskAnalysisTemplateServiceBuilder } from "../src/services/riskAnalysisService.js";
 
 export const { cleanup, readModelRepository, postgresDB, fileManager } =
   await setupTestContainersVitest(
@@ -41,11 +40,6 @@ export const eserviceTemplateService = eserviceTemplateServiceBuilder(
   postgresDB,
   readModelService,
   fileManager
-);
-
-export const riskAnalysisService = riskAnalysisTemplateServiceBuilder(
-  postgresDB,
-  readModelService
 );
 
 export const writeEServiceInEventstore = async (
